refactor(core): inline test helpers into test files
Move runInEntryContext helper from shared test-helpers.ts into each
test file that uses it. Test helpers belong with their tests, not in
separate utility files.

Verified: precommit-verify ✕
