+++ packages/core/src/modules/http/middleware/auth/auth.test.ts
@@ -1,9 +1,35 @@
 import type { Context } from 'hono';
 import { describe, expect, it } from 'vitest';
 
-import { runInEntryContext } from '../../internal/test-helpers';
+import { runInContext } from '../../../../kernel/internal/context-key';
+import { setBody } from '../../request/injection/body';
+import { setPathParams } from '../../request/injection/path-param';
+import { setHonoContext } from '../../request/request-context';
 
 import { currentRoles, currentUser, setUser } from './auth';
+
+type FormBody = Record<string, string | File | (string | File)[]>;
+
+type ParsedBody =
+  | { type: 'json'; val: unknown }
+  | { type: 'form'; val: FormBody }
+  | { type: 'text'; val: string }
+  | { type: 'none'; val: undefined };
+
+type TestEntryContext = {
+  honoContext: Context;
+  body?: ParsedBody;
+  pathParams?: Readonly<Record<string, string>>;
+};
+
+const runInEntryContext = <T>(ctx: TestEntryContext, fn: () => T): T => {
+  return runInContext(() => {
+    setHonoContext(ctx.honoContext);
+    setBody(ctx.body ?? { type: 'none', val: undefined });
+    setPathParams(ctx.pathParams ?? {});
+    return fn();
+  });
+};
 
 const createMockEntryContext = () => {
   const honoContext = {
