+++ packages/core/src/modules/http/request/entry-context.test.ts
@@ -1,10 +1,33 @@
 import type { Context } from 'hono';
 import { describe, expect, it } from 'vitest';
 
-import { runInEntryContext } from '../internal/test-helpers';
-import { body } from './injection/body';
-import { pathParam } from './injection/path-param';
-import { requestContext } from './request-context';
+import { runInContext } from '../../../kernel/internal/context-key';
+import { body, setBody } from './injection/body';
+import { pathParam, setPathParams } from './injection/path-param';
+import { requestContext, setHonoContext } from './request-context';
+
+type FormBody = Record<string, string | File | (string | File)[]>;
+
+type ParsedBody =
+  | { type: 'json'; val: unknown }
+  | { type: 'form'; val: FormBody }
+  | { type: 'text'; val: string }
+  | { type: 'none'; val: undefined };
+
+type TestEntryContext = {
+  honoContext: Context;
+  body?: ParsedBody;
+  pathParams?: Readonly<Record<string, string>>;
+};
+
+const runInEntryContext = <T>(ctx: TestEntryContext, fn: () => T): T => {
+  return runInContext(() => {
+    setHonoContext(ctx.honoContext);
+    setBody(ctx.body ?? { type: 'none', val: undefined });
+    setPathParams(ctx.pathParams ?? {});
+    return fn();
+  });
+};
 
 describe('entry-context', () => {
   it('returns the running context inside runInEntryContext', () => {
