+++ packages/core/src/modules/http/request/injection/get-context.test.ts
@@ -1,9 +1,34 @@
 import type { Context } from 'hono';
 import { describe, expect, it, vi } from 'vitest';
 
-import { runInEntryContext } from '../../internal/test-helpers';
-
+import { runInContext } from '../../../../kernel/internal/context-key';
+import { setHonoContext } from '../request-context';
+import { setBody } from './body';
 import { getContext, setContext } from './get-context';
+import { setPathParams } from './path-param';
+
+type FormBody = Record<string, string | File | (string | File)[]>;
+
+type ParsedBody =
+  | { type: 'json'; val: unknown }
+  | { type: 'form'; val: FormBody }
+  | { type: 'text'; val: string }
+  | { type: 'none'; val: undefined };
+
+type TestEntryContext = {
+  honoContext: Context;
+  body?: ParsedBody;
+  pathParams?: Readonly<Record<string, string>>;
+};
+
+const runInEntryContext = <T>(ctx: TestEntryContext, fn: () => T): T => {
+  return runInContext(() => {
+    setHonoContext(ctx.honoContext);
+    setBody(ctx.body ?? { type: 'none', val: undefined });
+    setPathParams(ctx.pathParams ?? {});
+    return fn();
+  });
+};
 
 declare module '@zeltjs/core' {
   interface RequestContextSchema {
